requirements.txt
Pillow
argparse
tqdm
